Break the client autodetect logic out of mod.ts
