+++ deno.json
@@ -2,7 +2,7 @@
   "name": "@cloudydeno/kubernetes-client",
   "version": "0.7.6",
   "imports": {
-    "@cloudydeno/stream-observables": "jsr:@cloudydeno/stream-observables@^1.4.1",
+    "@cloudydeno/stream-observables": "jsr:@cloudydeno/stream-observables@^1",
     "@std/path": "jsr:@std/path@^1",
     "@std/streams": "jsr:@std/streams@^1",
     "@std/yaml": "jsr:@std/yaml@^1"
@@ -12,7 +12,9 @@
     "./lib/kubeconfig.ts": "./lib/kubeconfig.ts",
     "./lib/reflector.ts": "./lib/reflector.ts",
     "./lib/stream-transformers.ts": "./lib/stream-transformers.ts",
+    "./transports": "./transports/mod.ts",
     "./transports/mod.ts": "./transports/mod.ts",
+    "./transports/autodetection.ts": "./transports/autodetection.ts",
     "./transports/via-kubeconfig.ts": "./transports/via-kubeconfig.ts",
     "./transports/via-kubectl-raw.ts": "./transports/via-kubectl-raw.ts",
     "./tunnel-beta/via-websocket.ts": "./tunnel-beta/via-websocket.ts",
